intelliwealth_model/financial_news/requirements.txt
flask
newspaper3k
nltk
beautifulsoup4
Pillow